requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.10.6
asyncpg==0.30.0
fastapi-cache2==0.2.2
redis==5.2.1
Faker==37.0.0